Migrate playercntl plugin to shared PluginUtilities
