+++ trunk/plugins/playercntl_plugin/Givecash.cpp
@@ -17,7 +17,7 @@
 #include <list>
 #include <set>
 
-#include "PluginUtilities.h"
+#include <PluginUtilities.h>
 #include "Main.h"
 
 #include <FLCoreServer.h>
