+++ trunk/plugins/playercntl_plugin/HyperJump.cpp
@@ -19,7 +19,7 @@
 #include <algorithm>
 
 #include "ZoneUtilities.h"
-#include "PluginUtilities.h"
+#include <PluginUtilities.h>
 #include "Main.h"
 #include "../../hookext_plugin/hookext_exports.h"
