+++ trunk/plugins/playercntl_plugin/Main.cpp
@@ -19,7 +19,7 @@
 #include <set>
 
 
-#include "PluginUtilities.h"
+#include <PluginUtilities.h>
 #include "ZoneUtilities.h"
 #include "CrashCatcher.h"
 #include "Main.h"
