+++ trunk/plugins/playercntl_plugin/PurchaseRestrictions.cpp
@@ -17,7 +17,7 @@
 #include <list>
 #include <set>
 
-#include "PluginUtilities.h"
+#include <PluginUtilities.h>
 #include "Main.h"
 
 namespace PurchaseRestrictions
